requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
ollama==0.1.7
lxml==4.9.3
feedparser==6.0.10 